Add support for skipping the registry in _testing
`load_response_set` and `load_response` can now take their relevant
object types directly. This allows use of the RegisteredResponse
interface even under parametrized tests and other scenarios where it's
not useful or desirable to register a ResponseSet for multiple tests
to use.

Convert the transport retry tests to use this behavior in order to
demonstrate.

Also, while doing this, reorganize `_testing` a little. The registry
module now contains all of the code for adding to and reading from the
registry, and the models have been moved into a dedicated module,
separate from the registry.
The main goal of this cleanup is to make the registry more readable,
as new and more nuanced behaviors are being added here.
